more convert

+++ src/convert.h
@@ -1,6 +1,6 @@
 #ifndef CONVERT_H
 #define CONVERT_H
 
-float convert_value (float in_value, const char* src, const char* dest);
+double convert_value (double in_value, const char* src, const char* dest);
 
 #endif
